% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chrono_strati_arg}
\alias{chrono_strati_arg}
\title{add chronostratigraphic bar to annotate an ggtree plot}
\usage{
chrono_strati_arg(
  start,
  offset_first = 0,
  offset_last = 0,
  entire = "Period",
  atom = "Age",
  chart = ggcsgb::Phanerozoic
)
}
\arguments{
\item{offset_first}{integer scalar. Add how many \code{entire} at the begining.}

\item{offset_last}{integer scalar. Add how many \code{entire} at the end.}

\item{entire}{string. One of \code{'Era'}, \code{'Period'}, \code{'Epoch'}, \code{'Age'}. The time unit to remain entire, since you have to keep it integrity if you want to add a label for that time unit.}

\item{atom}{string. Minimum time unit, only support \code{'Age'} now.}

\item{chart}{data.frame. The}
}
\value{
ggplot object chrono_strati_chart data to use, you usually needn't care it.
}
\description{
add chronostratigraphic bar to annotate an ggtree plot
}
\details{
Basic ggtree consists of a layer of horizontal lines and a layer of vertical lines, so \code{max(ggplot2::ggplot_build(plot)$data[[1]]$xend)} gives the time when the MRCA comes into being (previous \code{start} parameter).
}
