% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chrono_strati_axis}
\alias{chrono_strati_axis}
\title{add axis text}
\usage{
chrono_strati_axis(
  offset = 1,
  trim_left = T,
  squash = 0.015,
  present = T,
  angle = 90,
  size = 3,
  hjust = 0,
  ...
)
}
\arguments{
\item{offset}{numeric scalar. How much is the axis text higher than bar.}

\item{trim_left}{logical scalar. Whether to drop axis text which is before the MRCA}

\item{squash}{numeric scalar. Whether to remove some axis text if they are too close. If \code{0}, don't remove; else if the interval of two text is less than \code{squash*whole_axis_length}, one of the two is removed.}

\item{present}{numeric scalar. Whether to replace \code{0} with \code{present}}

\item{angle, size, hjust, ...}{other arguments passed on to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}
}
\description{
add axis text
}
