% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chrono_strati_bar}
\alias{chrono_strati_bar}
\title{add chronostrati bars to a plot}
\usage{
chrono_strati_bar(...)
}
\description{
add chronostrati bars implemented by \link[ggplot2:annotate]{ggplot2::annotate("rect")} to a ggplot object
}
