% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw_chrono_strati_bar}
\alias{draw_chrono_strati_bar}
\title{plot chronostratigraphic bar}
\usage{
draw_chrono_strati_bar(plot, data, start, height)
}
\arguments{
\item{plot}{ggplot2 object.}

\item{data}{data.frame. contains \code{color}, \code{begin}, \code{end}}

\item{start}{numeric scalar. The start of time, Mya.}

\item{height}{numeric scalar.}
}
\value{
ggplot2 object
}
\description{
plot chronostratigraphic bar
}
